feat: 기존 RequestProcessor 객체를 인터페이스 분리 후, 세션관련 추가 설정 기능을 Composite 하는 구현체 추가
