+++ src/main/java/webserver/http/config/WebServerConfig.java
@@ -1,7 +1,7 @@
 package webserver.http.config;
 
 import webserver.http.domain.controller.Controller;
-import webserver.http.domain.controller.RequestProcessor;
+import webserver.http.domain.controller.HttpRequestProcessor;
 import webserver.http.domain.controller.RootController;
 import webserver.http.domain.controller.StaticResourceController;
 import webserver.http.view.request.RequestReader;
@@ -19,7 +19,7 @@ public class WebServerConfig {
         return new ResponseWriter();
     }
 
-    public static RequestProcessor requestProcessor(List<Controller> controllers) {
+    public static HttpRequestProcessor requestProcessor(List<Controller> controllers) {
         List<Controller> totalControllers = new ArrayList<>(List.of(
                 new RootController(),
                 new StaticResourceController()
@@ -27,6 +27,6 @@ public class WebServerConfig {
 
         totalControllers.addAll(controllers);
 
-        return new RequestProcessor(totalControllers);
+        return new HttpRequestProcessor(totalControllers);
     }
 }
